requirements.txt
streamlit
pandas
requests
python-dotenv
plotly
streamlit-option-menu
scipy
matplotlib
finvizfinance
